BF_BOJ_16946_벽부수고이동하기4
https://www.acmicpc.net/problem/16946
used = set()
-Do not allow duplication.
-Unordered.
Because lists and tuples are ordered, the value of the data type can be obtained through indexing, but the set data type is unordered, so the value cannot be obtained through indexing. If you want to access the value stored in the set data type by indexing, you need to convert it to a list or tuple as follows.